+++ app/src/test/java/com/weighttrack/data/repo/FoodRepositoryTest.kt
@@ -37,7 +37,7 @@ class FoodRepositoryTest {
         foods = FoodRepository(
             database.foodDao(),
             OfflineFoodStore(ApplicationProvider.getApplicationContext()),
-            DeletionRecorder(database.deletionDao(), database.syncDao()),
+            DeletionRecorder(database, database.deletionDao(), database.syncDao()),
         )
     }